backend/app/security.py
import os
from cryptography.fernet import Fernet
from dotenv import load_dotenv

load_dotenv()

SECRET_KEY = os.getenv("SECRET_KEY", Fernet.generate_key())
cipher = Fernet(SECRET_KEY)

def encrypt_data(data: str) -> str:
    if not data:
        return data
    return cipher.encrypt(data.encode()).decode()

def decrypt_data(encrypted_data: str) -> str:
    if not encrypted_data:
        return encrypted_data
    try:
        return cipher.decrypt(encrypted_data.encode()).decode()
    except:
        return encrypted_data
