backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
bcrypt>=4.2.0
PyJWT>=2.9.0
python-multipart>=0.0.9
cryptography>=44.0.0
python-dotenv>=1.0.1
